ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
torch==2.3.0
torchaudio==2.3.0
transformers==4.41.2
bitsandbytes==0.43.1
accelerate==0.30.0
pydantic==2.7.1
pinecone-client==3.2.2
pymongo==4.7.2
cryptography==42.0.7
websockets==12.0
python-dotenv==1.0.1
# RVC v3 voice enhancement dependencies
faiss-cpu==1.8.0
praat-parselmouth==0.4.3
